# server/spec/openapi.yaml
openapi: "3.1.0"
info:
  title: Prompt Mint API
  version: "1.0.0"
  description: |
    Prompt Mint marketplace API — Express backend serving prompt listing,
    user management, governance, appeals, licensing, webhooks, versioning,
    notifications, analytics, and API key management.
  contact:
    name: Prompt Mint Team
  license:
    name: MIT
servers:
  - url: http://localhost:5000
    description: Local development server
  - url: https://api.promptmint.io
    description: Production server
paths:
...
  /api/prompts/{id}/versions:
    post:
      operationId: publishPromptVersion
      summary: Publish a new version of a prompt
      tags: [Prompts, Versions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                changeDescription:
                  type: string
      responses:
        "201":
          description: Version created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromptVersion"
        "404":
          $ref: "#/components/responses/NotFound"
    get:
      operationId: listPromptVersions
      summary: List all versions for a prompt
      tags: [Prompts, Versions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PromptVersion"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/prompts/{id}/versions/{versionIndex}:
    get:
      operationId: getPromptVersionDetail
      summary: Get a specific version of a prompt
      tags: [Prompts, Versions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: versionIndex
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Version detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromptVersion"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Buyer & Creator Prompt Endpoints ──────────────────────────────────────
...
  /api/versions/update:
    post:
      operationId: postPromptUpdate
      summary: Post a new content version for a prompt
      tags: [Versions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                promptId:
                  type: string
                content:
                  type: string
                changeDescription:
                  type: string
      responses:
        "200":
          description: Version created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromptVersion"
...
  /api/versions/purchase:
    post:
      operationId: recordPurchase
      summary: Record a purchase at the current version index
      tags: [Versions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                promptId:
                  type: string
                buyerWallet:
                  type: string
                transactionHash:
                  type: string
      responses:
        "200":
          description: Purchase recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  purchase:
                    $ref: "#/components/schemas/Purchase"
...
components:
  securitySchemes:
    adminToken:
      type: http
      scheme: bearer
      description: Admin API token from ADMIN_API_TOKEN env var
    apiKeyQuery:
      type: apiKey
      in: header
      name: X-Api-Key
      description: API key for programmatic access
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: string
...
  schemas:
...
    PromptVersion:
      type: object
      properties:
        _id:
          type: string
        promptId:
          type: string
        versionIndex:
          type: integer
        content:
          type: string
        changeDescription:
          type: string
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time

    # ── Marketplace Transaction ───────────────────────────────────────────────
...
    Purchase:
      type: object
      properties:
        _id:
          type: string
        promptId:
          type: string
        buyerWallet:
          type: string
        versionIndex:
          type: integer
        transactionHash:
          type: string
        createdAt:
          type: string
          format: date-time

    # ── Vote ──────────────────────────────────────────────────────────────────
...